@@ -1,4 +1,4 @@
-from typing import Callable, Union
+from collections.abc import Callable
 
 import gymnasium as gym
 import numpy as np
@@ -16,7 +16,7 @@ class ActionMasker(gym.Wrapper):
         or the name of such a method provided by the environment.
     """
 
-    def __init__(self, env: gym.Env, action_mask_fn: Union[str, Callable[[gym.Env], np.ndarray]]):
+    def __init__(self, env: gym.Env, action_mask_fn: str | Callable[[gym.Env], np.ndarray]):
         super().__init__(env)
 
         if isinstance(action_mask_fn, str):
