@@ -1,5 +1,3 @@
-from typing import Optional
-
 import gymnasium as gym
 import numpy as np
 import pytest
@@ -67,7 +65,7 @@ class DummyDictEnv(gym.Env):
         done = truncated = False
         return self.observation_space.sample(), reward, done, truncated, {}
 
-    def reset(self, *, seed: Optional[int] = None, options: Optional[dict] = None):
+    def reset(self, *, seed: int | None = None, options: dict | None = None):
         if seed is not None:
             self.observation_space.seed(seed)
         return self.observation_space.sample(), {}
