@@ -1,5 +1,3 @@
-from typing import Union
-
 import gymnasium as gym
 import numpy as np
 import pytest
@@ -80,7 +78,7 @@ def clone_tqc_batch_norm_stats(
     return (actor_bias, actor_running_mean, critic_bias, critic_running_mean, critic_target_bias, critic_target_running_mean)
 
 
-def clone_on_policy_batch_norm(model: Union[MaskablePPO]) -> (th.Tensor, th.Tensor):
+def clone_on_policy_batch_norm(model: MaskablePPO) -> (th.Tensor, th.Tensor):
     return clone_batch_norm_stats(model.policy.features_extractor.batch_norm)
 
 
@@ -121,7 +119,7 @@ def test_ppo_mask_train_eval_mode():
     batch_norm_stats_after = clone_on_policy_batch_norm(model)
 
     # No change in batch norm params
-    for param_before, param_after in zip(batch_norm_stats_before, batch_norm_stats_after):
+    for param_before, param_after in zip(batch_norm_stats_before, batch_norm_stats_after, strict=True):
         assert th.isclose(param_before, param_after).all()
 
 
@@ -235,7 +233,7 @@ def test_offpolicy_collect_rollout_batch_norm(model_class):
     batch_norm_stats_after = clone_helper(model)
 
     # No change in batch norm params
-    for param_before, param_after in zip(batch_norm_stats_before, batch_norm_stats_after):
+    for param_before, param_after in zip(batch_norm_stats_before, batch_norm_stats_after, strict=True):
         assert th.isclose(param_before, param_after).all()
 
 
@@ -274,5 +272,5 @@ def test_predict_with_dropout_batch_norm(model_class, env_id):
     batch_norm_stats_after = clone_helper(model)
 
     # No change in batch norm params
-    for param_before, param_after in zip(batch_norm_stats_before, batch_norm_stats_after):
+    for param_before, param_after in zip(batch_norm_stats_before, batch_norm_stats_after, strict=True):
         assert th.isclose(param_before, param_after).all()
